Move split function
